Solar_System_Project
This is my Solar System Project app that I submitted for the Mobile Apps Subject at Charles Sturt University (Ewan McDonnell 07/09/16).
Problems remaining - the touch and mouse interface were never implemented well. There is room for improvement.
I can't control the speed that the planets rotate at.
The planets rotations are supposed to be around the origin(0,0,0) and following a circle. This doesn't seem to work quite right, it looks like the Moon is floating off, which makes sense as it was set up to rotate around the earth.
Originally my idea was to be able to double-click on the objects of the solar system and open a link to the Wikipedia entry.